requirements.txt
pytaglib==0.3.6
pytest==6.0.1
colorama==0.4.3
